context/history/ref: seperate history logic
